Project

+++ PROJECT/Hotel/src/ee/taltech/iti0202/hotel/rooms/DoubleRoom.java
@@ -1,8 +1,10 @@
 package ee.taltech.iti0202.hotel.rooms;
 
+import java.math.BigDecimal;
+
 public class DoubleRoom extends Room {
 
-    public static final int AMOUNT = 80; // The price of the room
+    public static final BigDecimal AMOUNT = BigDecimal.valueOf(80); // The price of the room
 
     /**
      * Constructs a Double room which price is set to 80.
